Dropdown added in navbar
