+++ src/components/Navbarr/Navbarr.css
@@ -1,0 +1,26 @@
+.goog-te-combo{
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    background-color: #ffffff;
+    border: 1px solid #ced4da;
+    border-radius: 4px;
+    padding: 4px 8px;
+    font-size: 12px;
+    color: #6c757d;
+    min-width: 120px;
+    max-width: 140px;
+    height: 26px;
+    outline: none;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+    appearance: none; /* Remove default arrow on some browsers */
+    background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+    background-position: right 6px center;
+    background-repeat: no-repeat;
+    background-size: 14px;
+    padding-right: 26px;
+}
+
+.VIpgJd-ZVi9od-l4eHX-hSRGPd img{
+    display:inline;
+}
